# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Pourier API'
  description: "API REST pour la plateforme de vente de photos Pourier. Cette API permet aux utilisateurs d'acheter des photos, aux photographes de vendre leurs créations, et aux administrateurs de gérer la plateforme."
  contact:
    name: 'Pourier Support'
    email: contact@pourier.com
  license:
    name: Proprietary
    url: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur API Pourier'
paths:
...
  '/api/downloads/order/{order}':
    get:
      tags:
        - Downloads
      summary: 'Download all photos from order (ZIP)'
      description: 'Download all high-resolution photos from a completed order as a ZIP archive. User must be order owner.'
      operationId: getDownloadsOrder
      parameters:
        -
          name: order
          in: path
          description: 'Order UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'ZIP file download (streamed)'
          headers:
            Content-Disposition:
              description: 'Attachment filename'
              schema:
                type: string
                example: 'attachment; filename=order.zip'
          content:
            application/zip:
              schema:
                type: string
                format: binary
        403:
          description: 'Forbidden - order not completed or not owner'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Order is not completed yet'
                type: object
        404:
          description: 'No photos found in order'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'No photos found in this order'
                type: object
        500:
          description: 'Server error creating ZIP file'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Could not create zip file'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: "List user's orders"
      description: 'Get all orders for the authenticated user with pagination'
      operationId: getOrders
      parameters:
        -
          name: per_page
          in: query
          description: 'Number of orders per page'
          required: false
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        200:
          description: 'Orders retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Orders
      summary: 'Create new order'
      description: 'Create a new order with items from cart. Platform takes 20% commission.'
      operationId: storeOrders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - items
                - subtotal
                - total
                - payment_method
                - billing_email
                - billing_first_name
                - billing_last_name
                - billing_phone
              properties:
                items:
                  type: array
                  items:
                    properties:
                      photo_id:
                        type: string
                        format: uuid
                        example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
                      license_type:
                        type: string
                        enum:
                          - standard
                          - extended
                        example: standard
                    type: object
                  example:
                    -
                      photo_id: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
                      license_type: standard
                subtotal:
                  description: 'Sum of all item prices in FCFA'
                  type: number
                  format: float
                  example: 15000
                tax:
                  description: 'Tax amount (currently 0)'
                  type: number
                  format: float
                  example: 0
                discount:
                  description: 'Discount amount'
                  type: number
                  format: float
                  example: 0
                total:
                  description: 'Total amount in FCFA'
                  type: number
                  format: float
                  example: 15000
                payment_method:
                  type: string
                  enum:
                    - mobile_money
                    - card
                  example: mobile_money
                billing_email:
                  type: string
                  format: email
                  example: user@example.com
                billing_first_name:
                  type: string
                  example: Jean
                billing_last_name:
                  type: string
                  example: Dupont
                billing_phone:
                  description: 'Burkinabé phone format'
                  type: string
                  example: '+226 70 12 34 56'
              type: object
      responses:
        201:
          description: 'Order created successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Commande créée avec succès'
                  data:
                    $ref: '#/components/schemas/Order'
                type: object
        400:
          description: 'Order creation failed - photo unavailable or validation error'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Erreur lors de la création de la commande: La photo 'xyz' n'est plus disponible"
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        422:
          $ref: '#/components/responses/ValidationErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/orders/{order}':
    get:
      tags:
        - Orders
      summary: 'Get order details'
      description: 'Retrieve detailed information about a specific order (must be order owner)'
      operationId: getOrder
      parameters:
        -
          name: order
          in: path
          description: 'Order UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'Order details retrieved successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        403:
          description: 'Forbidden - user is not order owner'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
        404:
          $ref: '#/components/responses/NotFoundResponse'
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
  '/api/orders/{order}/pay':
    post:
      tags:
        - Orders
      summary: 'Initiate payment for order'
      description: 'Start payment process via CinetPay (Mobile Money or Card). Returns payment URL to redirect user.'
      operationId: payOrders
      parameters:
        -
          name: order
          in: path
          description: 'Order UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - payment_method
                - payment_provider
              properties:
                payment_method:
                  description: 'Payment method type'
                  type: string
                  enum:
                    - mobile_money
                    - card
                  example: mobile_money
                payment_provider:
                  description: 'CinetPay payment provider code'
                  type: string
                  enum:
                    - FLOOZ
                    - TMONEY
                    - MOOV
                    - CARD
                  example: FLOOZ
                phone:
                  description: 'Phone number for mobile money (required for mobile_money)'
                  type: string
                  example: '+22670123456'
              type: object
      responses:
        200:
          description: 'Payment initiated successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Paiement initié avec succès'
                  data:
                    properties:
                      payment_url:
                        description: 'URL to redirect user for payment'
                        type: string
                        format: url
                        example: 'https://client.cinetpay.com/payment/xyz123'
                      payment_token:
                        description: 'CinetPay payment token'
                        type: string
                        example: xyz123abc456
                    type: object
                type: object
        400:
          description: 'Payment initiation failed - order already processed or payment error'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Cette commande a déjà été traitée'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        422:
          $ref: '#/components/responses/ValidationErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/orders/{order}/status':
    get:
      tags:
        - Orders
      summary: 'Check order payment status'
      description: 'Check the current payment status of an order with CinetPay (must be order owner)'
      operationId: getOrdersStatus
      parameters:
        -
          name: order
          in: path
          description: 'Order UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'Payment status retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Statut de paiement récupéré'
                  data:
                    properties:
                      payment_status:
                        type: string
                        enum:
                          - pending
                          - completed
                          - failed
                          - cancelled
                        example: completed
                      order_status:
                        type: string
                        example: completed
                      transaction_id:
                        type: string
                        example: CP123456789
                    type: object
                type: object
        403:
          description: 'Forbidden - user is not order owner'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    ValidationErrorResponse:
      title: 'Validation Error Response'
      description: 'Réponse pour une erreur de validation'
      properties:
        message:
          type: string
          example: 'Les données fournies sont invalides'
        errors:
          description: 'Détails des erreurs de validation par champ'
          type: object
          example:
            email:
              - "L'email est invalide"
            password:
              - 'Le mot de passe est requis'
      type: object
    UnauthorizedResponse:
      title: 'Unauthorized Response'
      description: 'Réponse pour une authentification requise'
      properties:
        message:
          type: string
          example: 'Non authentifié'
      type: object
...
    PaginationMeta:
      title: 'Pagination Metadata'
      description: 'Métadonnées de pagination Laravel'
      properties:
        current_page:
          type: integer
          example: 1
        from:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 10
        per_page:
          type: integer
          example: 15
        to:
          type: integer
          example: 15
        total:
          type: integer
          example: 150
      type: object
    PaginationLinks:
      title: 'Pagination Links'
      description: 'Liens de pagination Laravel'
      properties:
        first:
          type: string
          example: 'http://api.pourier.com/api/photos?page=1'
        last:
          type: string
          example: 'http://api.pourier.com/api/photos?page=10'
        prev:
          type: string
          example: null
          nullable: true
        next:
          type: string
          example: 'http://api.pourier.com/api/photos?page=2'
      type: object
...
    OrderItem:
      title: 'Order Item'
      description: "Article d'une commande"
      properties:
        id:
          type: string
          format: uuid
        photo_id:
          type: string
          format: uuid
        license_type:
          type: string
          enum:
            - standard
            - extended
          example: standard
        price:
          type: number
          format: float
          example: 5000
        photo:
          $ref: '#/components/schemas/Photo'
      type: object
    Order:
      title: Order
      description: 'Commande complète'
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          example: ORD-20240115-001
        user_id:
          type: string
          format: uuid
        subtotal:
          type: number
          format: float
          example: 15000
        commission:
          description: '20% de commission'
          type: number
          format: float
          example: 3000
        total:
          type: number
          format: float
          example: 15000
        payment_status:
          type: string
          enum:
            - pending
            - completed
            - failed
            - refunded
          example: pending
        payment_method:
          type: string
          example: mobile_money
          nullable: true
        transaction_id:
          type: string
          example: CPAY-123456789
          nullable: true
        invoice_number:
          type: string
          example: INV-20240115-001
          nullable: true
        invoice_date:
          type: string
          format: date-time
          nullable: true
        invoice_path:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
  responses:
    UnauthorizedResponse:
      description: 'Unauthorized - Authentication required'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedResponse'
...
    NotFoundResponse:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundResponse'
    ValidationErrorResponse:
      description: 'Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter JWT Bearer token obtained from /api/auth/login endpoint'
...
security:
  -
    bearerAuth: []